Based on Development - Pre Task: JS -> TS conversion of existing codebase (PIC: Irfan Satria).
I have added Typescript and React types into the project and make sure that it can compile and run (npm start run)

I have also fixed typo in NavigatorBar, and changed the settings for TSconfig.json to allow implicit any and unstrict typing. We should be able to write normal JS in this project before putting types in it.

Thank you for your attention.